fix: admin scripts
